+++ application/src/main/java/org/opentripplanner/apis/gtfs/datafetchers/LegImpl.java
@@ -2,7 +2,6 @@ package org.opentripplanner.apis.gtfs.datafetchers;
 
 import graphql.schema.DataFetcher;
 import graphql.schema.DataFetchingEnvironment;
-import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
 import org.locationtech.jts.geom.Geometry;
@@ -13,6 +12,7 @@ import org.opentripplanner.apis.gtfs.mapping.LocalDateMapper;
 import org.opentripplanner.apis.gtfs.mapping.NumberMapper;
 import org.opentripplanner.apis.gtfs.mapping.PickDropMapper;
 import org.opentripplanner.apis.gtfs.mapping.RealtimeStateMapper;
+import org.opentripplanner.apis.gtfs.support.filter.StopArrivalByTypeFilter;
 import org.opentripplanner.ext.ridehailing.model.RideEstimate;
 import org.opentripplanner.ext.ridehailing.model.RideHailingLeg;
 import org.opentripplanner.framework.graphql.GraphQLUtils;
@@ -129,8 +129,8 @@ public class LegImpl implements GraphQLDataFetchers.GraphQLLeg {
     return environment -> getSource(environment).isInterlinedWithPreviousLeg();
   }
 
-  // TODO
   @Override
+  @Deprecated
   public DataFetcher<Boolean> intermediatePlace() {
     return environment -> false;
   }
@@ -142,12 +142,15 @@ public class LegImpl implements GraphQLDataFetchers.GraphQLLeg {
 
   @Override
   public DataFetcher<Iterable<Object>> intermediateStops() {
-    return environment -> {
-      List<StopArrival> intermediateStops = getSource(environment).listIntermediateStops();
+    return env -> {
+      var intermediateStops = getSource(env).listIntermediateStops();
       if (intermediateStops == null) {
         return null;
       }
-      return intermediateStops
+      var args = new GraphQLTypes.GraphQLLegIntermediateStopsArgs(env.getArguments());
+      var filter = new StopArrivalByTypeFilter(args.getGraphQLInclude());
+      return filter
+        .filter(intermediateStops)
         .stream()
         .map(intermediateStop -> intermediateStop.place.stop)
         .filter(Objects::nonNull)