+++ application/src/main/java/org/opentripplanner/routing/algorithm/raptoradapter/router/TransitRouter.java
@@ -258,7 +258,8 @@ public class TransitRouter {
       accessBuilder.withPreferences(p -> {
         p.withBike(b -> b.withRental(r -> r.withAllowArrivingInRentedVehicleAtDestination(false)));
         p.withCar(c -> c.withRental(r -> r.withAllowArrivingInRentedVehicleAtDestination(false)));
-        p.withScooter(s -> s.withRental(r -> r.withAllowArrivingInRentedVehicleAtDestination(false))
+        p.withScooter(s ->
+          s.withRental(r -> r.withAllowArrivingInRentedVehicleAtDestination(false))
         );
       });
     }