+++ application/src/main/java/org/opentripplanner/standalone/config/RouterConfig.java
@@ -8,6 +8,7 @@ import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.MissingNode;
 import java.io.Serializable;
 import java.util.List;
+import org.opentripplanner.apis.gtfs.GtfsApiParameters;
 import org.opentripplanner.ext.flex.FlexParameters;
 import org.opentripplanner.ext.ridehailing.RideHailingServiceParameters;
 import org.opentripplanner.ext.trias.config.TriasApiConfig;
@@ -20,6 +21,7 @@ import org.opentripplanner.standalone.config.routerconfig.TransitRoutingConfig;
 import org.opentripplanner.standalone.config.routerconfig.UpdatersConfig;
 import org.opentripplanner.standalone.config.routerconfig.VectorTileConfig;
 import org.opentripplanner.standalone.config.sandbox.FlexConfig;
+import org.opentripplanner.standalone.config.sandbox.GtfsApiConfig;
 import org.opentripplanner.standalone.config.sandbox.TransmodelAPIConfig;
 import org.opentripplanner.updater.UpdatersParameters;
 import org.slf4j.Logger;
@@ -50,6 +52,7 @@ public class RouterConfig implements Serializable {
   private final RideHailingServicesConfig rideHailingConfig;
   private final FlexConfig flexConfig;
   private final TransmodelAPIConfig transmodelApi;
+  private final GtfsApiConfig gtfsApi;
   private final VectorTileConfig vectorTileConfig;
   private final TriasApiParameters triasApiParameters;
 
@@ -69,6 +72,7 @@ public class RouterConfig implements Serializable {
 
     this.server = new ServerConfig("server", root);
     this.transmodelApi = new TransmodelAPIConfig("transmodelApi", root);
+    this.gtfsApi = new GtfsApiConfig("gtfsApi", root);
     var request = mapDefaultRouteRequest("routingDefaults", root);
     this.transitConfig = new TransitRoutingConfig("transit", root, request);
     this.routingRequestDefaults = request
@@ -140,6 +144,10 @@ public class RouterConfig implements Serializable {
 
   public TriasApiParameters triasApiParameters() {
     return triasApiParameters;
+  }
+
+  public GtfsApiParameters gtfsApiParameters() {
+    return gtfsApi;
   }
 
   public NodeAdapter asNodeAdapter() {