+++ application/src/test/java/org/opentripplanner/TestServerContext.java
@@ -106,6 +106,7 @@ public class TestServerContext {
       routerConfig.transitTuningConfig(),
       transitService,
       routerConfig.triasApiParameters(),
+      routerConfig.gtfsApiParameters(),
       routerConfig.vectorTileConfig(),
       createVehicleParkingService(),
       createVehicleRentalService(),